create category and converter pages
